Title: 두 수의 합, Time: 3020ms, Memory: 176MB, Status: Time Limit Exceed - CodeTree

+++ 240519/두 수의 합/sum-of-two-num.java
@@ -6,21 +6,22 @@ public class Main {
 
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
-        Map<Integer,Integer> map = new HashMap<>();
+        Map<Long,Integer> map = new HashMap<>();
         String[]StrArr = br.readLine().split(" ");
 
         int n = Integer.parseInt(StrArr[0]);
-        int k = Integer.parseInt(StrArr[1]);
+        Long k = Long.parseLong(StrArr[1]);
 
         String[]arr = br.readLine().split(" ");
-        int[]intArr = new int[n];
+        Long[]intArr = new Long[n];
         for(int i =0;i <n;i++){
-            intArr[i] = Integer.parseInt(arr[i]);
+            intArr[i] = Long.parseLong(arr[i]);
         }
 
         for(int i=0;i<n-1;i++){
             for(int j = i+1;j<n;j++){
-                int sum = intArr[i]+intArr[j];
+                Long sum = intArr[i]+intArr[j];
+                // Long sum = Long.valueOf(intArr[i])+Long.valueOf(intArr[j]);
                 map.put(sum,map.getOrDefault(sum,0)+1);
             }
         }
